# server/openapi.yaml
openapi: 3.1.0
info:
  title: EV2 Clandestinoz API
  version: 0.1.0
  description: |
    API del sistema de gestión de nightclub EV2 Clandestinoz.

    **Este archivo es el contrato entre backend, web y apps móviles.** Cualquier cambio
    de ruta o de forma de respuesta se refleja aquí en el mismo commit que lo introduce.

    ## Autenticación
    Salvo las rutas marcadas como públicas, toda petición requiere
    `Authorization: Bearer <access_token>`. El token de acceso dura 15 minutos; se renueva
    con `POST /api/auth/refresh` usando el refresh token, que es de **un solo uso**
    (al canjearlo se revoca y se emite uno nuevo).

    ## Errores
    Todos los errores tienen la misma forma (`ErrorResponse`). Los códigos usados son
    400 validación, 401 sin autenticar, 403 sin permiso, 404 no existe, 409 conflicto de
    estado, 422 regla de negocio, 429 límite de peticiones, 500 interno, 501 no implementado.

    ## Dinero
    Los importes son cadenas decimales con dos decimales (`"1500.00"`) para no perder
    precisión, y siempre van acompañados de `currency` (`MXN` o `USD`). Los precios se
    calculan **siempre en el servidor**.

    ## Precios de reservación
    Una mesa se reserva por **toda la noche**, que corresponde a un evento del calendario.
    El precio es `base de la zona + (personas extra x precio del boleto de esa noche) +
    extras`. Cada zona incluye un número de boletos y admite un tope de extras (por
...

...

    ## Flirt
...

    ## Empleados y dinero
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/nightclubs/{nightclubId}:
    get:
      operationId: getNightclubsByNightclubId
      tags: [Club]
      summary: Perfil del club
      parameters: [{ $ref: '#/components/parameters/NightclubId' }]
      responses:
        '200':
          description: Perfil
          content:
            application/json:
              schema:
                type: object
                properties:
                  nightclub: { $ref: '#/components/schemas/Nightclub' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/nightclubs/{nightclubId}/tables:
    get:
      operationId: getNightclubsByNightclubIdTables
      tags: [Mesas]
      summary: Mapa de mesas con sus ocupantes
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - name: section
          in: query
          schema: { type: string, example: ZONA DIAMANTE }
        - name: floor
          in: query
          description: EV2 Clandestinoz tiene dos plantas.
          schema: { type: string, enum: [baja, alta, ambas] }
        - name: status
          in: query
          schema:
            type: string
            enum: [available, occupied, reserved, blocked, cleaning]
      responses:
        '200':
          description: Mesas
          content:
            application/json:
              schema:
                type: object
                properties:
                  tables:
                    type: array
                    items: { $ref: '#/components/schemas/Table' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/nightclubs/{nightclubId}/floor-plan:
    get:
      operationId: getNightclubsByNightclubIdFloorPlan
      tags: [Mesas]
      summary: Plano completo del local
      description: |
        Todo lo necesario para dibujar el mapa: mesas con sus coordenadas y ocupación
        actual, las áreas que no son mesas (barra, pista, cabina del DJ, entrada, baños)
        y el tamaño del lienzo. EV2 Clandestinoz tiene dos plantas: sin filtro se
        devuelven ambas.
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - name: floor
          in: query
          schema: { type: string, enum: [baja, alta, ambas] }
      responses:
        '200':
          description: Plano
          content:
            application/json:
              schema:
                type: object
                properties:
                  canvas:
                    type: [object, 'null']
                    properties:
                      width: { type: integer }
                      height: { type: integer }
                  floors: { type: array, items: { type: string } }
                  tables:
                    type: array
                    items:
                      allOf:
                        - { $ref: '#/components/schemas/Table' }
                        - type: object
                          properties:
                            seated: { type: integer, description: Personas sentadas ahora mismo. }
                  landmarks:
                    type: array
                    items: { $ref: '#/components/schemas/Landmark' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/nightclubs/{nightclubId}/tables/stats:
    get:
      operationId: getNightclubsByNightclubIdTablesStats
      tags: [Mesas]
      summary: Ocupación del local
      description: |
        Requiere personal de piso (`waiter`, `hostess`, `bartender`, `manager`).
        Devuelve el total del club y el desglose por piso y por zona, con el porcentaje
        de mesas ocupadas y el de lugares usados sobre el aforo.
      parameters: [{ $ref: '#/components/parameters/NightclubId' }]
      responses:
        '200':
          description: Ocupación
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { $ref: '#/components/schemas/OccupancyBucket' }
                  floors:
                    type: array
                    items:
                      allOf:
                        - type: object
                          properties:
                            floor: { type: string }
                        - { $ref: '#/components/schemas/OccupancyBucket' }
                        - type: object
                          properties:
                            sections:
                              type: array
                              items:
                                allOf:
                                  - type: object
                                    properties:
                                      section: { type: string }
                                  - { $ref: '#/components/schemas/OccupancyBucket' }
                  generated_at: { type: string, format: date-time }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/nightclubs/{nightclubId}/tables/layout:
    put:
      operationId: putNightclubsByNightclubIdTablesLayout
      tags: [Mesas]
      summary: Guardar el plano de mesas
      description: |
        Requiere rol `manager`. Actualiza en bloque coordenadas, radio, sección,
        capacidad y estado. Se usa desde el editor visual del portal de gerencia.
      parameters: [{ $ref: '#/components/parameters/NightclubId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tables]
              properties:
                tables:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    type: object
                    required: [id]
                    properties:
                      id: { type: string, format: uuid }
                      x: { type: number }
                      y: { type: number }
                      radius: { type: number, exclusiveMinimum: 0 }
                      section: { type: string }
                      capacity: { type: integer, minimum: 1 }
                      status:
                        type: string
                        enum: [available, occupied, reserved, blocked, cleaning]
                      bottle_service: { type: boolean }
      responses:
        '200':
          description: Mesas actualizadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: integer }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/nightclubs/{nightclubId}/tables/{tableId}/seat:
    post:
      operationId: postNightclubsByNightclubIdTablesByTableIdSeat
      tags: [Mesas]
      summary: Sentarse en una mesa
      description: |
        El usuario deja automáticamente cualquier otra mesa. Rechaza con 409 si la mesa
        está llena, bloqueada o en limpieza.
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - { $ref: '#/components/parameters/TableId' }
      responses:
        '201':
          description: Sentado
          content:
            application/json:
              schema:
                type: object
                properties:
                  seated: { type: boolean }
                  table_id: { type: string, format: uuid }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/nightclubs/{nightclubId}/tables/{tableId}/release:
    post:
      operationId: postNightclubsByNightclubIdTablesByTableIdRelease
      tags: [Mesas]
      summary: Levantar a alguien de una mesa
      description: |
        Solo personal (`waiter`, `hostess`, `manager`, `admin`), y `user_id` es
        obligatorio: dice a quién se levanta. El cliente no se sienta ni se levanta solo
        desde que la entrada se controla en la puerta.
      parameters:
        - { $ref: '#/components/parameters/NightclubId' }
        - { $ref: '#/components/parameters/TableId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        '200':
          description: Liberado
          content:
            application/json:
              schema:
                type: object
                properties:
                  released: { type: boolean }
                  table_id: { type: string, format: uuid }
                  remaining: { type: integer, description: Personas que siguen en la mesa }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso de 15 minutos obtenido en login o refresh.
    agentKey:
      type: apiKey
      in: header
      name: X-Agent-Key
      description: |
        Llave propia del agente local del POS. Autentica una maquina, no una persona:
        no hereda permisos de ningun usuario y solo sirve en las rutas /api/pos/agent/*.
  parameters:
    NightclubId:
      name: nightclubId
      in: path
      required: true
      description: Debe coincidir con el club del usuario autenticado; si no, 403.
      schema: { type: string, format: uuid }
...
    TableId:
      name: tableId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Falta el token o no es válido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Sin permiso (rol insuficiente u otro club)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: El recurso no existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflicto con el estado actual del recurso
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict,
                     unprocessable, too_many_requests, internal_error, not_implemented]
            message: { type: string }
            details:
              description: Presente en errores de validación y de disponibilidad.
              type: array
              items: { type: object, additionalProperties: true }
            request_id:
              type: string
              description: Identificador de la petición, útil para buscar en los registros.
...
    Currency:
      type: string
      enum: [MXN, USD]
...
    NightclubPublic:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        city: { type: string }
        country: { type: string }
        timezone: { type: string, example: America/Hermosillo }
        currency_default: { $ref: '#/components/schemas/Currency' }
    Nightclub:
      allOf:
        - { $ref: '#/components/schemas/NightclubPublic' }
        - type: object
          properties:
            address: { type: string }
            phone: { type: string }
            capacity: { type: integer }
            settings: { type: object, additionalProperties: true }
...
    Table:
      type: object
      properties:
        id: { type: string, format: uuid }
        code:
          type: string
          description: |
            Identificador único dentro del club. Normalmente es el número de la mesa;
            lleva prefijo de zona donde el número se repite (V16 en VIP Elevado frente
            a A16 en Zona Azul).
          example: V16
        table_number:
          type: [integer, 'null']
          description: El número con el que el personal se refiere a la mesa en voz alta.
          example: 16
        name: { type: [string, 'null'] }
        floor: { type: string, enum: [baja, alta, ambas] }
        color: { type: [string, 'null'], description: Color de la zona en el mapa., example: '#FF1493' }
        section: { type: string, example: ZONA DIAMANTE }
        type: { type: string, enum: [standard, vip, premium, bar_top, booth] }
        capacity: { type: integer }
        x: { type: string, description: Coordenada en el plano. }
        y: { type: string }
        radius: { type: string }
        status: { type: string, enum: [available, occupied, reserved, blocked, cleaning] }
        bottle_service: { type: boolean }
        occupants:
          type: array
          items:
            type: object
            properties:
              user_id: { type: string, format: uuid }
              display_name: { type: string }
              seated_at: { type: string, format: date-time }
    Landmark:
      description: Elemento del mapa que no es una mesa.
      type: object
      properties:
        code: { type: string, example: dj_booth }
        name: { type: string, example: DJ BOOTH }
        type:
          type: string
          enum: [bar, dance_area, dj, vip, entrance, restroom, stage, other]
        description: { type: [string, 'null'] }
        floor: { type: string, enum: [baja, alta, ambas] }
        x: { type: [string, 'null'] }
        y: { type: [string, 'null'] }
        width: { type: [string, 'null'] }
        height: { type: [string, 'null'] }
    OccupancyBucket:
      type: object
      properties:
        tables: { type: integer }
        occupied: { type: integer }
        available: { type: integer }
        reserved: { type: integer }
        out_of_service: { type: integer, description: Bloqueadas o en limpieza. }
        seats: { type: integer, description: Aforo sentado. }
        guests: { type: integer, description: Personas sentadas ahora mismo. }
        tables_occupied_pct: { type: integer }
        seats_used_pct: { type: integer }
...